separate facet ordering structures, add fallback strategy in case of missing values
